requirements/base.txt
Django==5.0.6
mysqlclient==2.2.4
Pillow==10.3.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-allauth==0.63.3
django-filter==24.2
djangorestframework==3.15.1
python-decouple==3.8
whitenoise==6.7.0
